requirements.txt
streamlit
python-dotenv
openai-agents
pillow
